Eliminated compile warning

+++ OldSoar/trunk/soar/kernel/osupport.c
@@ -406,7 +406,8 @@ void calculate_support_for_instantiation_preferences (instantiation *inst) {
 			       } else if ( current_agent(o_support_calculation_type) == 4 &&
 												 (rhs_value_is_reteloc(act->id)) &&
 												 w->value == get_symbol_from_rete_loc( (byte)rhs_value_to_reteloc_levels_up( act->id ),
-																													 rhs_value_to_reteloc_field_num( act->id ), inst->rete_token, w )) {
+																					   (byte)rhs_value_to_reteloc_field_num( act->id ),
+																					   inst->rete_token, w )) {
 							 op_elab = TRUE;
